refactoring and different removal strategy
