+++ github_actions_runner_launcher.go
@@ -5,57 +5,31 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"os/signal"
 	"path/filepath"
-	"regexp"
 	"sync"
 	"syscall"
 
 	"github.com/google/go-github/v31/github"
-	"github.com/k0kubun/pp"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/oauth2"
-	"gopkg.in/yaml.v2"
 )
 
 // Version is incremented using bump2version
 const Version = "0.1.1"
 
-// RunnerEnvironment ...
-type RunnerEnvironment struct {
-	RepoURL       string `yaml:"REPO_URL"`
-	AccessToken   string `yaml:"ACCESS_TOKEN"`
-	RunnerName    string `yaml:"RUNNER_NAME"`
-	RunnerToken   string `yaml:"RUNNER_TOKEN"`
-	RunnerWorkdir string `yaml:"RUNNER_WORKDIR"`
-	IsOrgRunner   string `yaml:"ORG_RUNNER"`
-	OrgName       string `yaml:"ORG_NAME"`
-	Labels        string `yaml:"LABELS"`
-}
-
-// RunnerConfig ...
-type RunnerConfig struct {
-	Environment RunnerEnvironment `yaml:"environment"`
-}
-
-// LaunchConfig ...
-type LaunchConfig struct {
-	Services map[string]RunnerConfig `yaml:"services"`
-}
-
 // Launcher ...
 type Launcher struct {
-	Config           LaunchConfig
-	RunnerVersion    string
-	RunnerArch       string
-	ForceReconfigure bool
-	ForceRemoveExisting bool
-	configPath       string
-	apiClient        *github.Client
-	aptMux 			 sync.Mutex
+	Config         LaunchConfig
+	RunnerVersion  string
+	RunnerArch     string
+	Reconfigure    bool
+	RemoveExisting bool
+	configPath     string
+	apiClient      *github.Client
+	aptMux         sync.Mutex
 }
 
 // NewWithConfig ...
@@ -112,33 +86,26 @@ func (l *Launcher) Run(run bool) (err error) {
 
 		go func(rConf RunnerConfig, rToken string) {
 			wg.Add(1)
-			if err := l.configureRunner(rLog, rConf, rToken); err != nil {
+			if err := l.ConfigureRunner(rLog, rConf, rToken); err != nil {
 				rLog.Error(err)
 			}
 			if run {
 				l.startRunner(ctx, rLog, &wg, rConf, rToken)
 			}
 			wg.Done()
-		}(runnerConfig, runnerToken) 
+		}(runnerConfig, runnerToken)
 	}
 	wg.Wait()
 	return nil
 }
 
 func (l *Launcher) startRunner(ctx context.Context, rLog *log.Entry, wg *sync.WaitGroup, runnerConfig RunnerConfig, runnerToken string) {
-	// wg.Add(1)
-	// defer wg.Done()
-
 	workDir, err := l.GetWorkDirForRunner(runnerConfig)
 	if err != nil {
 		rLog.Error(err)
 		return
 	}
 
-	if err := l.configureRunner(rLog, runnerConfig, runnerToken); err != nil {
-		rLog.Error(err)
-		return
-	}
 	cmd := exec.CommandContext(ctx, filepath.Join(workDir, "./run.sh"))
 	cmd.Dir = workDir
 
@@ -175,86 +142,29 @@ func (l *Launcher) startRunner(ctx context.Context, rLog *log.Entry, wg *sync.Wa
 	}
 }
 
-// ParseConfigFile ...
-func (l *Launcher) configureRunner(rLog *log.Entry, runner RunnerConfig, runnerToken string) error {
+// ConfigureRunner ...
+func (l *Launcher) ConfigureRunner(rLog *log.Entry, runner RunnerConfig, runnerToken string) error {
 	workDir, err := l.GetWorkDirForRunner(runner)
 	if err != nil {
 		return err
 	}
-	if err := l.prepareRunnerFiles(rLog, runner); err != nil {
+	if err := l.PrepareRunnerFiles(rLog, runner); err != nil {
 		return fmt.Errorf("Failed to prepare runner: %v", err)
 	}
 	cmd := exec.Command(filepath.Join(workDir, "./config.sh"), "--url", runner.Environment.RepoURL, "--token", runnerToken, "--name", runner.Environment.RunnerName, "--work", workDir, "--labels", runner.Environment.Labels, "--unattended", "--replace")
 	cmd.Dir = workDir
-	if out, err := cmd.CombinedOutput(); l.ForceReconfigure && err != nil {
+	out, err := cmd.CombinedOutput()
+	rLog.Debug(string(out))
+	if l.Reconfigure && err != nil {
 		rLog.Error(cmd.String())
 		rLog.Error(string(out))
-		return fmt.Errorf("Failed to configure runner and ForceReconfigure option is set: %v", err)
+		return fmt.Errorf("Failed to configure runner and Reconfigure option is set: %v", err)
 	}
 	return nil
 }
 
-// ParseConfigFile ...
-func (l *Launcher) ParseConfigFile(file string) error {
-	data, err := ioutil.ReadFile(file)
-	if err != nil {
-		return fmt.Errorf("Failed to read config file %v: %v", file, err)
-	}
-	if err := yaml.Unmarshal(data, &l.Config); err != nil {
-		return fmt.Errorf("Failed to parse config: %v", err)
-	}
-	if log.IsLevelEnabled(log.DebugLevel) {
-		pp.Print(l.Config)
-	}
-	l.configPath = file
-	log.Info("Parsed config")
-	return nil
-}
-
-// ObtainRunnerToken ...
-func (l *Launcher) ObtainRunnerToken(ctx context.Context, runner RunnerConfig, accessToken string) (string, *github.Timestamp, error) {
-	if l.apiClient == nil {
-		l.apiClient = createGitHubAPIClient(ctx, accessToken)
-	}
-	r := regexp.MustCompile(`^.*github\.com/(?P<Acc>.*)/(?P<Repo>.*)$`)
-	matches := r.FindStringSubmatch(runner.Environment.RepoURL)
-	if len(matches) != 3 {
-		return "", nil, fmt.Errorf("Failed to extract github account and repo name from \"%s\"", runner.Environment.RepoURL)
-	}
-	acc, repo := matches[1], matches[2]
-	log.Debugf("acc=%s, repo=%s", acc, repo)
-	if acc == "" || repo == "" {
-		return "", nil, fmt.Errorf("Failed to extract github account and repo name from \"%s\" (acc=%s, repo=%s)", runner.Environment.RepoURL, acc, repo)
-	}
-
-	runners, _, err := l.apiClient.Actions.ListRunners(ctx, acc, repo, &github.ListOptions{
-		Page: 0,
-		PerPage: 100,
-	})
-	if err != nil {
-		log.Warnf("Failed to check for any existing runners: %v", err)
-	} else if l.ForceRemoveExisting {
-		log.Infof("%d existing runners will be removed", runners.TotalCount)
-		for _, runner := range runners.Runners {
-			log.Infof("Removing runner %s (%d) [os=%s, status=%s]", runner.GetName(), runner.GetID(), runner.GetOS(), runner.GetStatus())
-			if _, err := l.apiClient.Actions.RemoveRunner(ctx, acc, repo, runner.GetID()); err != nil {
-				log.Warnf("Failed to remove runner %d: %v", runner.GetID(), err)
-			}
-		}
-	}
-	token, response, err := l.apiClient.Actions.CreateRegistrationToken(ctx, acc, repo)
-	log.Debugf("runner_token=%v", token)
-	if err != nil || token.Token == nil {
-		return "", nil, fmt.Errorf("Failed to obtain runner token from the GitHub API: %v", err)
-	}
-	log.Debugf("%d of %d GitHub API requests left. Will be reset on %v", response.Rate.Remaining, response.Rate.Limit, response.Rate.Reset)
-	if response.Rate.Remaining < 50 {
-		log.Warningf("Only %d of %d GitHub API requests left. Will be reset on %v", response.Rate.Remaining, response.Rate.Limit, response.Rate.Reset)
-	}
-	return *token.Token, token.ExpiresAt, nil
-}
-
-func createGitHubAPIClient(ctx context.Context, accessToken string) *github.Client {
+// CreateGitHubAPIClient ...
+func CreateGitHubAPIClient(ctx context.Context, accessToken string) *github.Client {
 	tokenSource := &TokenSource{
 		AccessToken: accessToken,
 	}
